obs(bootstrap): measure the payload the publisher actually ships against its byte budget
The PR-time budget suite can only see tier MEMBERSHIP. It looks each key up in
the frozen 2026-08-21 ledger via tierPayloadBytesFromLedger, so a key whose
production row count doubles moves no number it reads, and its 5% per-key growth
check runs against synthetic representative fixtures rather than production
volume.

Between the 2026-08-27 and 2026-08-28 scheduled DebugBear runs the slow tier
grew +223,845 B decoded on desktop (86545576 -> 86622879) and +133,874 B on
mobile (86545531 -> 86622827) with no membership change at all. Mobile crossed
its 3,000 ms abort, truncated, and re-triggered 130,357 B of the per-panel
fallbacks #7048 removed: list-fire-detections, list-natural-events and
sanctionsPressure. The budget suite was green at 9/9 throughout.

The publisher is the only place that sees the bytes actually shipped, so it now
evaluates the real ledger against the same ceilings and emits a distinct
[bootstrap-budget] warning line an alert can key on.

This is an observation, not a gate. #7045's stop conditions forbid putting
mutable production values into a pull-request pass/fail check, and refusing to
publish over a size regression would take the dashboard down to prevent it being
slow — so a breach warns and the publish proceeds. A test pins that.

The frozen ledger, ceilings, manifest and evaluator move to
shared/bootstrap-payload-budget.js because scripts/ may not import from tests/.
The fixture re-exports them unchanged, so there is still one copy of the
captured numbers and the budget suite is untouched at 9/9.

Refs #7288
